refactor(orders): flatten routing to use search params and remove type path segment (#113282)
