--- app/[tenant]/[workspace]/(subapps)/orders/[type]/layout.tsx
@@ -1,25 +1,0 @@
-import {notFound} from 'next/navigation';
-
-// ---- LOCAL IMPORTS ---- //
-import {ORDER_TAB_ITEMS} from '@/subapps/orders/common/constants/orders';
-
-export default async function Layout(props: {
-  params: Promise<{
-    type: string;
-    tenant: string;
-    workspace: string;
-  }>;
-  children: React.ReactNode;
-}) {
-  const params = await props.params;
-
-  const {children} = props;
-
-  const {type} = params;
-
-  if (!ORDER_TAB_ITEMS.some(item => item.href === type)) {
-    return notFound();
-  }
-
-  return <div>{children}</div>;
-}
