+++ constants/index.ts
@@ -95,7 +95,6 @@ export const SUBAPP_WITH_ROLES = [
  * SUBAPP PAGE
  */
 export const SUBAPP_PAGE = {
-  orders: '/ongoing',
   completed: 'completed',
   archived: '/archived',
   checkout: `${SUBAPP_CODES.shop}/cart/checkout`,
